# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zenythium Fitness API",
        "description": "API для управления тренировками, упражнениями и планами тренировок",
        "contact": {
            "email": "support@zenythium.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Основной сервер API"
        }
    ],
    "paths": {
...
        "/api/v1/muscle-groups": {
            "get": {
                "tags": [
                    "Muscle Groups"
                ],
                "summary": "Получение списка групп мышц",
                "description": "Возвращает пагинированный список групп мышц с возможностью фильтрации. Доступно как аутентифицированным, так и неаутентифицированным пользователям",
                "operationId": "97173868df12197f1826ec228cdf4bb8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Фильтр по названию группы мышц",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "грудь"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Группы мышц успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Группы мышц успешно получены"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Muscle Groups"
                ],
                "summary": "Создание новой группы мышц",
                "description": "Создает новую группу мышц. Доступно всем пользователям",
                "operationId": "aff74f66f5a67d12e728614cd5c29a47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название группы мышц",
                                        "type": "string",
                                        "example": "Грудь"
                                    },
                                    "description": {
                                        "description": "Описание группы мышц",
                                        "type": "string",
                                        "example": "Мышцы грудной клетки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Группа мышц успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц успешно создана"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/muscle-groups/{muscleGroup}": {
            "get": {
                "tags": [
                    "Muscle Groups"
                ],
                "summary": "Получение конкретной группы мышц",
                "description": "Возвращает детальную информацию о группе мышц по ID. Доступно как аутентифицированным, так и неаутентифицированным пользователям",
                "operationId": "8eac1d427981e58407f7176bce1d7022",
                "parameters": [
                    {
                        "name": "muscleGroup",
                        "in": "path",
                        "description": "ID группы мышц",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Группа мышц успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц успешно получена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Группа мышц не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Muscle Groups"
                ],
                "summary": "Обновление группы мышц",
                "description": "Обновляет информацию о существующей группе мышц",
                "operationId": "c545615ff39eb61bba7c614eb1114f9e",
                "parameters": [
                    {
                        "name": "muscleGroup",
                        "in": "path",
                        "description": "ID группы мышц",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Название группы мышц",
                                        "type": "string",
                                        "example": "Грудь"
                                    },
                                    "description": {
                                        "description": "Описание группы мышц",
                                        "type": "string",
                                        "example": "Мышцы грудной клетки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Группа мышц успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц успешно обновлена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Группа мышц не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Muscle Groups"
                ],
                "summary": "Удаление группы мышц",
                "description": "Удаляет группу мышц и все связанные с ней упражнения",
                "operationId": "d22ae5c9a85f5e9747683bc937bfd517",
                "parameters": [
                    {
                        "name": "muscleGroup",
                        "in": "path",
                        "description": "ID группы мышц",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Группа мышц успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц успешно удалена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Группа мышц не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Группа мышц не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}